Move per-task restorers to dexjoco.tasks.state_restorers
The replay script shrinks from 567 to 349 lines and now imports
`restore_initial_state` / `has_restorer` from the new module. The
public entry point is unchanged; this is purely a relocation, plus
two fidelity tweaks:

- `_restore_click_mouse` now also updates `display_root.qpos[2]` to
  `_display_root_z0 + restored_delta_h`. The monitor is a free joint
  with gravity; reset() pins its z to the originally-sampled delta_h,
  so after the restorer overwrites delta_h the monitor would float or
  bury. Across visual seeds this leaked into the contact solver and
  compounded to ~1e-1 drift by the end of an episode.
- `_restore_click_mouse` also re-attaches the mousepad to the recorded
  mouse_xy at the correct table z (read from the table site rather
  than `_table_body_z0 + delta_h`), so the contact graph matches.

With these in place, record-vs-replay diff is 0.0 on all 11 tasks /
12 example demos across different visual seeds (full episode length).
