+++ .gitignore
@@ -7,6 +7,7 @@ __pycache__/
 /outputs/
 /checkpoints
 /datasets
+/_replay_test/
 
 teleoperation/GeoRT/checkpoint/**/*.pth
 !teleoperation/GeoRT/checkpoint/fk_model_allegro_left.pth
